perf: ⚡️ Use Float32Array and avoid duplicated slice insertion
Use a Float32Array and fill it using the cornerstone Int16 data to avoid
recalculation. Also prevent multiple loadImageData calls on the same
imageDataObject from overloading each other.

+++ src/lib/getImageData.js
@@ -55,7 +55,7 @@ export default function getImageData(imageIds, displaySetInstanceUid) {
       }
 
     case 16:
-      pixelArray = new Int16Array(xVoxels * yVoxels * zVoxels);
+      pixelArray = new Float32Array(xVoxels * yVoxels * zVoxels);
 
       break;
   }
